Add files via upload
This is a basic level Hostel Management System that help to understand basic functions of Warden and a Student/Hostellite. Server is made up of Node js that take input from user side ,which is made of Html,CSS,Js , and go to realted file fetch data and send it to front page. this is all functionality.